refactor: extract shared pipeline from wasm.rs for native testing
Extract the core evaluation pipeline from wasm.rs into wasm_pipeline.rs
so that native tests exercise the exact same code path (including
parse_unit) as WASM users. The wasm.rs module becomes a thin wrapper
that converts PipelineError into the JSON ErrorInfo envelope.

Add 10 unit tests covering JSON/YAML/text output, parse errors, unknown
formats, divergent programs, nested blocks, lists, booleans, and
expression evaluation.
